[ALPS02008661] TOI: remove build warnings
[Detail]
printk arguments type mismatch
[Solution]
1. Use %lu for unsigned long types
2. Use type casting for non-prmitive types

[Feature] Instant Power On

MTK-Commit-Id: b5065cef957d9563f764c6430a8b05daea4898b6

CR-Id: ALPS02008661

+++ kernel/power/tuxonice_prepare_image.c
@@ -901,9 +901,9 @@ static void flag_image_pages(int atomic_copy)
 
 	if (!atomic_copy)
 		toi_message(TOI_EAT_MEMORY, TOI_MEDIUM, 0,
-			    "Count data pages: Set1 (%d) + Set2 (%d) + Nosave (%ld) + NumFree (%d) = %d.\n",
+			    "Count data pages: Set1 (%lu) + Set2 (%lu) + Nosave (%lu) + NumFree (%d) = %lu.\n",
 			    pagedir1.size, pagedir2.size, num_nosave, num_free,
-			    pagedir1.size + pagedir2.size + num_nosave + num_free);
+			    (unsigned long) (pagedir1.size + pagedir2.size + num_nosave + num_free));
 }
 
 void toi_recalculate_image_contents(int atomic_copy)
